src/components/ErrorBoundary/ErrorBoundary.tsx
import React, { Component } from "react";
import Error from "../../interfaces/Error.interface";

class ErrorBoundary extends Component<{}, Error> {
  constructor(props: {}) {
    super(props);
    this.state = { hasError: false };
  }

  static getDerivedStateFromError(error: {}) {
    return { hasError: true };
  }

  render() {
    if (this.state.hasError) {
      // You can render any custom fallback UI
      return <h1>Something went wrong.</h1>;
    }

    return this.props.children;
  }
}

export default ErrorBoundary;
